backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
asyncpg
alembic
pydantic
passlib[bcrypt]
python-jose
python-multipart
pydantic-settings
email-validator
bcrypt==3.2.0